refactor FE to a few components
