+++ apps/opik-frontend/src/components/pages-shared/automations/AddEditRuleDialog/RuleFilteringSection.tsx
@@ -1,0 +1,285 @@
+import React, { useCallback, useMemo } from "react";
+import { UseFormReturn } from "react-hook-form";
+import { Plus } from "lucide-react";
+import uniqid from "uniqid";
+
+import { Button } from "@/components/ui/button";
+import { Label } from "@/components/ui/label";
+import { FormField, FormItem } from "@/components/ui/form";
+import {
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+} from "@/components/ui/accordion";
+import { Filter } from "@/types/filters";
+import { COLUMN_TYPE, ColumnData } from "@/types/shared";
+import {
+  COLUMN_METADATA_ID,
+  COLUMN_FEEDBACK_SCORES_ID,
+  COLUMN_CUSTOM_ID,
+} from "@/types/shared";
+import { TRACE_DATA_TYPE } from "@/hooks/useTracesOrSpansList";
+import { CUSTOM_FILTER_VALIDATION_REGEXP } from "@/constants/filters";
+import { createFilter } from "@/lib/filters";
+import FiltersContent from "@/components/shared/FiltersContent/FiltersContent";
+import TracesOrSpansPathsAutocomplete from "@/components/pages-shared/traces/TracesOrSpansPathsAutocomplete/TracesOrSpansPathsAutocomplete";
+import TracesOrSpansFeedbackScoresSelect from "@/components/pages-shared/traces/TracesOrSpansFeedbackScoresSelect/TracesOrSpansFeedbackScoresSelect";
+import SliderInputControl from "@/components/shared/SliderInputControl/SliderInputControl";
+import { EVALUATORS_RULE_SCOPE } from "@/types/automations";
+import { EvaluationRuleFormType } from "./schema";
+import ExplainerIcon from "@/components/shared/ExplainerIcon/ExplainerIcon";
+import { Description } from "@/components/ui/description";
+
+const AUTOMATION_RULE_FILTER_COLUMNS: ColumnData<TRACE_DATA_TYPE>[] = [
+  {
+    id: "id",
+    label: "ID",
+    type: COLUMN_TYPE.string,
+  },
+  {
+    id: "name",
+    label: "Name",
+    type: COLUMN_TYPE.string,
+  },
+  {
+    id: "start_time",
+    label: "Start time",
+    type: COLUMN_TYPE.time,
+  },
+  {
+    id: "end_time",
+    label: "End time",
+    type: COLUMN_TYPE.time,
+  },
+  {
+    id: "input",
+    label: "Input",
+    type: COLUMN_TYPE.string,
+  },
+  {
+    id: "output",
+    label: "Output",
+    type: COLUMN_TYPE.string,
+  },
+  {
+    id: "duration",
+    label: "Duration",
+    type: COLUMN_TYPE.duration,
+  },
+  {
+    id: COLUMN_METADATA_ID,
+    label: "Metadata",
+    type: COLUMN_TYPE.dictionary,
+  },
+  {
+    id: "tags",
+    label: "Tags",
+    type: COLUMN_TYPE.list,
+  },
+  {
+    id: "thread_id",
+    label: "Thread ID",
+    type: COLUMN_TYPE.string,
+  },
+  {
+    id: COLUMN_FEEDBACK_SCORES_ID,
+    label: "Feedback scores",
+    type: COLUMN_TYPE.numberDictionary,
+  },
+  {
+    id: COLUMN_CUSTOM_ID,
+    label: "Custom filter",
+    type: COLUMN_TYPE.dictionary,
+  },
+];
+
+const DEFAULT_SAMPLING_RATE = 1;
+
+interface RuleFilteringSectionProps {
+  form: UseFormReturn<EvaluationRuleFormType>;
+  projectId: string;
+}
+
+const RuleFilteringSection: React.FC<RuleFilteringSectionProps> = ({
+  form,
+  projectId,
+}) => {
+  const scope = form.watch("scope");
+  const filtersConfig = useMemo(
+    () => ({
+      rowsMap: {
+        [COLUMN_METADATA_ID]: {
+          keyComponent: TracesOrSpansPathsAutocomplete as React.FC<unknown> & {
+            placeholder: string;
+            value: string;
+            onValueChange: (value: string) => void;
+          },
+          keyComponentProps: {
+            rootKeys: ["metadata"],
+            projectId,
+            type: TRACE_DATA_TYPE.traces,
+            placeholder: "key",
+            excludeRoot: true,
+          },
+        },
+        [COLUMN_CUSTOM_ID]: {
+          keyComponent: TracesOrSpansPathsAutocomplete as React.FC<unknown> & {
+            placeholder: string;
+            value: string;
+            onValueChange: (value: string) => void;
+          },
+          keyComponentProps: {
+            rootKeys: ["input", "output"],
+            projectId,
+            type: TRACE_DATA_TYPE.traces,
+            placeholder: "key",
+            excludeRoot: false,
+          },
+          validateFilter: (filter: Filter) => {
+            if (
+              filter.key &&
+              filter.value &&
+              !CUSTOM_FILTER_VALIDATION_REGEXP.test(filter.key)
+            ) {
+              return `Key is invalid, it should begin with "input", or "output" and follow this format: "input.[PATH]" For example: "input.message" `;
+            }
+          },
+        },
+        [COLUMN_FEEDBACK_SCORES_ID]: {
+          keyComponent:
+            TracesOrSpansFeedbackScoresSelect as React.FC<unknown> & {
+              placeholder: string;
+              value: string;
+              onValueChange: (value: string) => void;
+            },
+          keyComponentProps: {
+            projectId,
+            type: TRACE_DATA_TYPE.traces,
+            placeholder: "Select score",
+          },
+        },
+      },
+    }),
+    [projectId],
+  );
+
+  const handleAddFilter = useCallback(() => {
+    const currentFilters = form.getValues("filters");
+    const newFilter = {
+      ...createFilter(),
+      id: uniqid(),
+    };
+    form.setValue("filters", [...currentFilters, newFilter]);
+  }, [form]);
+
+  const setFilters = useCallback(
+    (filtersOrUpdater: Filter[] | ((prev: Filter[]) => Filter[])) => {
+      if (typeof filtersOrUpdater === "function") {
+        const currentFilters = form.getValues("filters");
+        const updatedFilters = filtersOrUpdater(currentFilters);
+        form.setValue("filters", updatedFilters);
+      } else {
+        form.setValue("filters", filtersOrUpdater);
+      }
+    },
+    [form],
+  );
+
+  const isTraceScope = scope === EVALUATORS_RULE_SCOPE.trace;
+
+  return (
+    <Accordion
+      type="single"
+      collapsible
+      className="-mb-4 w-full border-t border-border"
+    >
+      <AccordionItem value="filtering-sampling" className="border-none">
+        <AccordionTrigger className="px-3 py-2 hover:no-underline">
+          <div className="flex items-center gap-1">
+            <Label className="text-sm font-medium">
+              {isTraceScope ? "Filtering & Sampling" : "Sampling"}
+            </Label>
+            <ExplainerIcon
+              className="mt-0.5"
+              description={
+                isTraceScope
+                  ? "Apply filters and sampling to select which traces will be evaluated by this rule"
+                  : "Use sampling rate to control how frequently this rule is applied to threads"
+              }
+            />
+          </div>
+        </AccordionTrigger>
+        <AccordionContent className="px-3 pb-3">
+          <div className="mb-8 space-y-4">
+            <Description>
+              Use sampling rate to control how frequently this rule is applied.
+              {isTraceScope &&
+                " For trace scope, you can also add filters to select specific traces based on their properties."}{" "}
+              If nothing is defined, the rule will evaluate all{" "}
+              {scope === EVALUATORS_RULE_SCOPE.trace ? "traces" : "threads"}.
+            </Description>
+
+            {isTraceScope && (
+              <FormField
+                control={form.control}
+                name="filters"
+                render={({ field }) => (
+                  <FormItem>
+                    <div className="space-y-3">
+                      <Label className="text-sm font-medium">Filters</Label>
+
+                      {field.value.length > 0 && (
+                        <FiltersContent
+                          filters={field.value}
+                          setFilters={setFilters}
+                          columns={AUTOMATION_RULE_FILTER_COLUMNS}
+                          config={filtersConfig}
+                          className="py-0"
+                        />
+                      )}
+
+                      <div className="pt-1">
+                        <Button
+                          type="button"
+                          variant="outline"
+                          size="sm"
+                          onClick={handleAddFilter}
+                          className="w-fit"
+                        >
+                          <Plus className="mr-1 size-3.5" />
+                          Add filter
+                        </Button>
+                      </div>
+                    </div>
+                  </FormItem>
+                )}
+              />
+            )}
+
+            {/* Sampling Rate */}
+            <FormField
+              control={form.control}
+              name="samplingRate"
+              render={({ field }) => (
+                <SliderInputControl
+                  min={0}
+                  max={1}
+                  step={0.01}
+                  defaultValue={DEFAULT_SAMPLING_RATE}
+                  value={field.value}
+                  onChange={field.onChange}
+                  id="sampling_rate"
+                  label="Sampling rate"
+                  tooltip="Percentage of traces to evaluate"
+                />
+              )}
+            />
+          </div>
+        </AccordionContent>
+      </AccordionItem>
+    </Accordion>
+  );
+};
+
+export default RuleFilteringSection;
