+++ apps/opik-frontend/src/components/pages-shared/automations/AddEditRuleDialog/helpers.ts
@@ -1,8 +1,11 @@
+import uniqid from "uniqid";
+
 import {
   EVALUATORS_RULE_SCOPE,
   EVALUATORS_RULE_TYPE,
   UI_EVALUATORS_RULE_TYPE,
 } from "@/types/automations";
+import { Filter } from "@/types/filters";
 
 export const getUIRuleType = (ruleType: EVALUATORS_RULE_TYPE) =>
   ({
@@ -37,3 +40,20 @@ export const getBackendRuleType = (
         EVALUATORS_RULE_TYPE.thread_python_code,
     },
   })[scope][uiType];
+
+export const normalizeFilters = (filters?: Filter[]): Filter[] => {
+  if (!filters || filters.length === 0) return [];
+
+  return filters.map(
+    (filter) =>
+      ({
+        id: filter.id || uniqid(),
+        field: filter.field || "",
+        type: filter.type || "string",
+        operator: filter.operator || "",
+        key: filter.key || "",
+        value: filter.value || "",
+        error: filter.error || "",
+      }) as Filter,
+  );
+};
