+++ apps/opik-frontend/src/components/shared/FiltersButton/FiltersButton.tsx
@@ -1,6 +1,4 @@
-import React, { useCallback, useEffect, useMemo, useState } from "react";
-import filter from "lodash/filter";
-import uniq from "lodash/uniq";
+import React, { useCallback, useEffect, useState } from "react";
 import { Filter as FilterIcon, Plus } from "lucide-react";
 import {
   Popover,
@@ -9,10 +7,10 @@ import {
 } from "@/components/ui/popover";
 import { Button, ButtonProps } from "@/components/ui/button";
 import { Separator } from "@/components/ui/separator";
-import { Filter, FilterRowConfig, Filters } from "@/types/filters";
+import { FilterRowConfig, Filters } from "@/types/filters";
 import { ColumnData } from "@/types/shared";
 import { createFilter, isFilterValid } from "@/lib/filters";
-import FilterRow from "@/components/shared/FiltersButton/FilterRow";
+import FiltersContent from "@/components/shared/FiltersContent/FiltersContent";
 import useDeepMemo from "@/hooks/useDeepMemo";
 import { cn } from "@/lib/utils";
 
@@ -46,8 +44,16 @@ const FiltersButton = <TColumnData,>({
   const isIconLayout = layout === "icon";
 
   const validFilters = useDeepMemo(() => {
-    return filter(filters, isFilterValid);
+    return filters.filter(isFilterValid);
   }, [filters]);
+
+  const onClearAll = useCallback(() => {
+    setFilters([]);
+  }, [setFilters]);
+
+  const onAddFilter = useCallback(() => {
+    setFilters((prev) => [...prev, createFilter()]);
+  }, [setFilters]);
 
   useEffect(() => {
     if (!open) {
@@ -62,62 +68,6 @@ const FiltersButton = <TColumnData,>({
   useEffect(() => {
     onChange(validFilters);
   }, [validFilters, onChange]);
-
-  const clearHandler = useCallback(() => {
-    setFilters([]);
-  }, []);
-
-  const addHandler = useCallback(() => {
-    setFilters((state) => [...state, createFilter()]);
-  }, []);
-
-  const onRemoveRow = useCallback((id: string) => {
-    setFilters((state) => filter(state, (f) => f.id !== id));
-  }, []);
-
-  const onChangeRow = useCallback((updateFilter: Filter) => {
-    setFilters((state) =>
-      state.map((f) => (updateFilter.id === f.id ? updateFilter : f)),
-    );
-  }, []);
-
-  const disabledColumns = useMemo(() => {
-    const disposableColumns = columns
-      .filter((c) => c.disposable)
-      .map((c) => c.id);
-
-    if (!disposableColumns.length) {
-      return undefined;
-    }
-
-    const columnsWIthFilters = uniq(filters.map((f) => f.field));
-
-    return disposableColumns.filter((c) => columnsWIthFilters.includes(c));
-  }, [filters, columns]);
-
-  const getConfig = useCallback(
-    (field: string) => config?.rowsMap[field],
-    [config],
-  );
-
-  const renderFilters = () => {
-    return filters.map((filter, index) => {
-      const prefix = index === 0 ? "Where" : "And";
-
-      return (
-        <FilterRow
-          key={filter.id}
-          columns={columns}
-          getConfig={getConfig}
-          disabledColumns={disabledColumns}
-          filter={filter}
-          prefix={prefix}
-          onRemove={onRemoveRow}
-          onChange={onChangeRow}
-        />
-      );
-    });
-  };
 
   return (
     <Popover onOpenChange={setOpen} open={open}>
@@ -149,19 +99,21 @@ const FiltersButton = <TColumnData,>({
               variant="ghost"
               size="sm"
               className="-mr-2.5"
-              onClick={clearHandler}
+              onClick={onClearAll}
             >
               Clear all
             </Button>
           </div>
           <Separator />
-          <div className="-mr-1 max-h-[50vh] overflow-y-auto overflow-x-hidden py-4">
-            <table className="table-auto">
-              <tbody>{renderFilters()}</tbody>
-            </table>
-          </div>
+          <FiltersContent<TColumnData>
+            filters={filters}
+            setFilters={setFilters}
+            columns={columns}
+            config={config}
+            className="-mr-1 max-h-[50vh]"
+          />
           <div className="flex items-center">
-            <Button variant="secondary" onClick={addHandler}>
+            <Button variant="secondary" onClick={onAddFilter}>
               <Plus className="mr-2 size-4" />
               Add filter
             </Button>
