+++ apps/opik-frontend/src/components/shared/FiltersContent/OperatorSelector.tsx
