+++ apps/opik-frontend/src/components/shared/FiltersContent/rows/ListRow.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Filter } from "@/types/filters";
-import OperatorSelector from "@/components/shared/FiltersButton/OperatorSelector";
+import OperatorSelector from "@/components/shared/FiltersContent/OperatorSelector";
 import DebounceInput from "@/components/shared/DebounceInput/DebounceInput";
 import { DEFAULT_OPERATORS, OPERATORS_MAP } from "@/constants/filters";
 import { COLUMN_TYPE } from "@/types/shared";
